requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
tqdm==4.67.1
urllib3==2.5.0
